pyproject.toml
[project]
name = "eid-card"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=0.8.0",
    "emoji>=2.10.0",
    "fpdf2>=2.7.7",
    "jinja2>=3.0.0",
    "litellm>=1.20.0",
    "loguru>=0.7.2",
    "openrouter>=0.3.0",
    "pexels-api>=1.0.1",
    "pillow>=8.3.1",
    "playwright>=1.41.0",
    "pydantic>=2.6.0",
    "python-dotenv>=0.19.0",
    "python-multipart>=0.0.6",
    "requests>=2.26.0",
    "streamlit>=1.24.0",
    "swarms>=0.3.0",
    "weasyprint>=60.2",
]
